+++ tests/tests_pytorch/trainer/connectors/test_data_connector.py
@@ -1,4 +1,4 @@
-# Copyright The PyTorch Lightning team.
+# Copyright The Lightning AI team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -90,7 +90,6 @@ def test_replace_distributed_sampler(tmpdir, mode):
             return [self.create_dataset()] * self._numbers_test_dataloaders
 
     model = TestModel(2, mode)
-    model.test_epoch_end = None
 
     trainer = Trainer(
         default_root_dir=tmpdir,